auto_test_pipeline/requirements.txt
diffusers
peft
packaging
matplotlib
seaborn
scikit-learn
sentence-transformers
xformers
torch
transformers
pillow
numpy
